+++ web/src/js/ducks/_options_gen.ts
@@ -79,6 +79,7 @@ export interface OptionsState {
     stream_large_bodies: string | undefined;
     strip_ech: boolean;
     tcp_hosts: string[];
+    tcp_timeout: number;
     termlog_verbosity: string;
     tls_ecdh_curve_client: string | undefined;
     tls_ecdh_curve_server: string | undefined;
@@ -185,6 +186,7 @@ export const defaultState: OptionsState = {
     stream_large_bodies: undefined,
     strip_ech: true,
     tcp_hosts: [],
+    tcp_timeout: 600,
     termlog_verbosity: "info",
     tls_ecdh_curve_client: undefined,
     tls_ecdh_curve_server: undefined,